+++ data/Duke.txt
@@ -1,3 +1,3 @@
 T | 1 | sleep
 T | 1 | sleep
-D | 0 | sleep | 2020-11-11 1200
+D | 1 | sleep | 2020-11-11 1200
