+++ src/main/java/duke/Event.java
@@ -1,34 +1,40 @@
+package duke;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
-    private String Dtime;
+
+public class Event extends Task {
+    private String schedule;
     private LocalDateTime localDtime;
+    private LocalTime hour;
 
-    public Deadline(String description, String s) {
+    public Event(String description, String s) {
         super(description);
-        this.Dtime = s; // accept "yyyy-mm-dd kkm" format
+        this.schedule = s; // accept "yyyy-mm-dd kkmm" format
         this.localDtime = LocalDateTime.parse(s.substring(1), DateTimeFormatter.ofPattern("uuuu-MM-dd kkmm"));
-
     }
 
     @Override
     public String toString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("uuuu-MM-dd kkmm");
         String formatDateTime = this.localDtime.format(formatter);
-        return "[D]" + super.toString() + " (by:" + formatDateTime + ")";
+
+        return "[E]" + super.toString() + " (at:" + formatDateTime + ")";
     }
+
 
     @Override
     public String toStringConvert(){
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("uuuu-MM-dd kkmm");
         String formatDateTime = this.localDtime.format(formatter);
+
         if(this.isCompleted()) {
-            return "D | 1 | " +  this.getString() + "| " + formatDateTime;
+            return "E | 1 | " +  this.getString() + "| " + formatDateTime;
         } else {
-            return "D | 0 | " +  this.getString() + "| " + formatDateTime;
+            return "E | 0 | " +  this.getString() + "| " + formatDateTime;
         }
     }
+
 
 }
