+++ src/main/java/duke/InputChecker.java
@@ -1,3 +1,4 @@
+package duke;
 public class InputChecker {
     public String input;
